fix tasks
